pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "masi"
version = "0.1.0"
description = "Proposal-aligned research scaffold for MASI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Pradyun Devarakonda" }
]
dependencies = []

[project.optional-dependencies]
recommender = [
  "numpy>=2.4.3",
  "Pillow>=10.0.0",
  "torch>=2.10.0",
  "transformers>=4.45.0"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
